feat(DONE dialog): Add time input function to 'DONE property'
